build(core): support scala native
