+++ core/jvm-native/src/main/scala/zio/logging/internal/FileWriter.scala
