+++ core/jvm-native/src/main/scala/zio/logging/internal/WriterProvider.scala
