+++ core/shared/src/test/scala/zio/logging/LogFormatPatternSpec.scala
