+++ examples/core/jvm/src/main/resources/logger.conf
