+++ examples/core/jvm/src/main/scala/zio/logging/api/http/ApiDomain.scala
