+++ examples/core/jvm/src/main/scala/zio/logging/example/ConfigurableLoggerApp.scala
