+++ examples/core/jvm/src/main/scala/zio/logging/example/ConsoleJsonApp.scala
