+++ examples/core/jvm/src/main/scala/zio/logging/example/FileApp.scala
