+++ examples/core/jvm/src/test/scala/zio/logging/example/PingServiceSpec.scala
