+++ examples/core/shared/src/main/scala/zio/logging/ConfigurableLogger.scala
