+++ examples/core/shared/src/test/scala/zio/logging/example/LoggingSpec.scala
